Feat/dlesage/v0.15 (#252)
adding code for release v0.15

+++ python/mod_cvcuda/OpInpaint.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2023-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 2023-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -44,7 +44,7 @@ public:
 
         size_t payloadSize() const
         {
-            return m_maxShape.w * m_maxShape.h * m_maxBatchSize;
+            return static_cast<size_t>(m_maxShape.w) * m_maxShape.h * m_maxBatchSize;
         }
 
     private:
